Распределил код задания 8.1 на 2 модуля, убрал лишние глобальные переменные
