+++ js/viewing-cards.js
@@ -1,38 +1,10 @@
+import { fillBigPicture } from './fill-big-picture';
 import { isEscape } from './utils';
 
-const pictures = document.querySelector('.pictures');
-const bigPicture = document.querySelector('.big-picture');
-const bigPictureClose = bigPicture.querySelector('.big-picture__cancel');
-const commentsList = bigPicture.querySelector('.social__comments');
-const commentTemplate = document.querySelector('#comment-item').content.querySelector('.social__comment');
+const pictures = document.querySelector('.pictures'); // Список с миниатюрами
+const bigPicture = document.querySelector('.big-picture'); // Большое фото
 
-const fillBigPicture = (card) => {
-  const image = bigPicture.querySelector('.big-picture__img img');
-  image.src = card.url;
-  const likesCounter = bigPicture.querySelector('.likes-count');
-  likesCounter.textContent = card.likes;
-  const commentsCurrentCounter = bigPicture.querySelector('.social__comment-shown-count');
-  commentsCurrentCounter.textContent = card.comments.length;
-  const commentsCounter = bigPicture.querySelector('.social__comment-total-count');
-  commentsCounter.textContent = card.comments.length;
-
-  const ourFragment = document.createDocumentFragment();
-  card.comments.forEach((commentData) => {
-    const newCommentElement = commentTemplate.cloneNode(true);
-    const avatar = newCommentElement.querySelector('.social__picture');
-    avatar.src = commentData.avatar;
-    avatar.alt = commentData.name;
-    const text = newCommentElement.querySelector('.social__text');
-    text.textContent = commentData.message;
-    ourFragment.appendChild(newCommentElement);
-  });
-  commentsList.innerHTML = '';
-  commentsList.appendChild(ourFragment);
-
-  const description = document.querySelector('.social__caption');
-  description.textContent = card.description;
-};
-
+// Обработчик события по клавише
 const onDocumentKeyDown = (evt) => {
   if (isEscape(evt)) {
     evt.preventDefault();
@@ -41,28 +13,33 @@ const onDocumentKeyDown = (evt) => {
   }
 };
 
+
+// Обработчик события по клику на миниатюру
 const openBigPicture = (evt, cardList) => {
   if (evt.target.matches('.picture__img')) {
     bigPicture.classList.remove('hidden');
 
     document.body.classList.add('modal-open');
 
-    document.addEventListener('keydown', onDocumentKeyDown);
+    document.addEventListener('keydown', onDocumentKeyDown); // Добавляет обработчик события по клавише
 
-    const pictureLink = evt.target.parentElement;
-    const id = Number(pictureLink.dataset.id);
+    const pictureLink = evt.target.parentElement; // Находит родителя миниатюры
+    const id = Number(pictureLink.dataset.id); // Находит id из дата-атрибута родителя
 
-    const card = cardList.find((item) => item.id === id);
+    const card = cardList.find((item) => item.id === id); // Ищет по id карточку миниатюры в списке карточек
 
-    fillBigPicture(card);
+    fillBigPicture(card, bigPicture); // Вызывает функцию заполнения фотокарточки
   }
 };
 
+
+// Инициализатор со слушателем события по клику
 const initialize = (cardList) => {
-  pictures.addEventListener('click', (evt) => {
+  const bigPictureClose = bigPicture.querySelector('.big-picture__cancel');
+  pictures.addEventListener('click', (evt) => { // Слушатель события по клику на список
     openBigPicture(evt, cardList);
   });
-  bigPictureClose.addEventListener('click', () => {
+  bigPictureClose.addEventListener('click', () => { // Слушатель события по клику на кнопку закрытия
     bigPicture.classList.add('hidden');
     document.body.classList.remove('modal-open');
     document.removeEventListener('keydown', onDocumentKeyDown);
@@ -70,3 +47,10 @@ const initialize = (cardList) => {
 };
 
 export {initialize};
+
+/*
+Написать программу, которая показывает только 5 первых комментариев из списка.
+По нажатию кнопки с классом comments-loader должны показываться следующие 5 комментариев.
+Количество показанных комментариев должно совпадать с тем, которое в данный момент отображается.
+
+*/
